fix: SOS GPS from FIELD_TELEMETRY, tracker restore on restart, audio ANR
1. parseSosLocation now checks FIELD_TELEMETRY before falling back to
   text regex. Extracted as shared top-level function used by both
   MessageCollector (trail + notifications) and MessagingScreen (UI).
   Fixes missing breadcrumb trail for Sideband-compatible clients.

2. SosActiveTracker is restored from DB when MessageCollector starts
   (not just in Application.onCreate), so service disconnect/reconnect
   within a session no longer loses SOS indicators.

3. audioRecorder.start() runs on Dispatchers.IO instead of Main to
   avoid ANR from MediaRecorder.prepare() blocking the main thread.
